add datastore for contact and additional comment features
contact form now connected to datastore to store email messages.
There are two new features added for comments. One is the button
that can delete all comments. The other is user can input the number
of comments they want to see.

Blog post commenting features are in the process of being added, but
is currently incomplete.